requirements.txt
# Core dependencies
anthropic>=0.18.0
sentence-transformers>=2.2.0
pandas>=2.0.0
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional adapters (install as needed)
# langchain>=0.1.0
# llama-index>=0.10.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
